client location interpolate

+++ Project/Source/Project/Public/NetworkingThread.h
@@ -26,10 +26,10 @@ public:
 
     bool IsRunning = true;
 	bool IsConnected = false;
-    int32 CycleTime{};
+    double CycleTime{};
 
 private: 
-   // FString IPAddress = "192.168.0.8";
+    //FString IPAddress = "192.168.0.8";
     FString IPAddress = "127.0.0.1";
     int32 PortNumber = 9000;
     FSocket* Socket;
